# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ant-design v5 接口demo",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/admin_users/ids_destroy": {
            "delete": {
                "tags": [
                    "用户"
                ],
                "summary": "删除用户",
                "description": "删除用户",
                "operationId": "App\\Http\\Controllers\\AdminUsersController::destroy",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "用户id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/menus": {
            "get": {
                "tags": [
                    "菜单"
                ],
                "summary": "菜单列表",
                "description": "菜单列表",
                "operationId": "App\\Http\\Controllers\\MenusController::index",
                "responses": {
                    "200": {
                        "description": "角色列表"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "菜单"
                ],
                "summary": "添加菜单",
                "description": "添加菜单",
                "operationId": "App\\Http\\Controllers\\MenusController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "菜单名",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "上级菜单id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功"
                    },
                    "422": {
                        "description": "错误的凭证响应"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/menus/{menu}": {
            "post": {
                "tags": [
                    "菜单"
                ],
                "summary": "添加菜单",
                "description": "添加菜单",
                "operationId": "App\\Http\\Controllers\\MenusController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "菜单id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "菜单名",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "上级菜单id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功"
                    },
                    "422": {
                        "description": "错误的凭证响应"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/permissions": {
            "get": {
                "tags": [
                    "权限"
                ],
                "summary": "权限列表",
                "description": "权限列表",
                "operationId": "App\\Http\\Controllers\\PermissionsController::index",
                "responses": {
                    "200": {
                        "description": "权限列表"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "权限"
                ],
                "summary": "添加权限",
                "description": "添加权限",
                "operationId": "App\\Http\\Controllers\\PermissionsController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "权限名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "权限别称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功"
                    },
                    "422": {
                        "description": "错误的凭证响应"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/permissions/{permission}": {
            "patch": {
                "tags": [
                    "权限"
                ],
                "summary": "修改权限",
                "description": "修改权限",
                "operationId": "App\\Http\\Controllers\\PermissionsController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "权限id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "权限名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "权限别称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/permissions/ids_destroy": {
            "delete": {
                "tags": [
                    "权限"
                ],
                "summary": "删除权限",
                "description": "删除权限",
                "operationId": "App\\Http\\Controllers\\PermissionsController::destroy",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "权限id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/permissions/role_by_permissions/{roleId?}": {
            "get": {
                "tags": [
                    "权限"
                ],
                "summary": "根据角色获取拥有权限",
                "description": "根据角色获取拥有权限",
                "operationId": "App\\Http\\Controllers\\PermissionsController::getRoleByPermissions",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "query",
                        "description": "角色id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "权限列表"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/roles/ids_destroy": {
            "delete": {
                "tags": [
                    "角色"
                ],
                "summary": "删除角色",
                "description": "删除角色",
                "operationId": "App\\Http\\Controllers\\RolesController::destroy",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "角色id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}